requirements.txt
git+https://github.com/adsabs/ADSMicroserviceUtils.git@v1.2.1
Flask-Limiter==1.4
iiif-prezi==0.3.0
SQLAlchemy-Utils==0.38.2
Flask-Compress==1.12
Jinja2==2.11.3
MarkupSafe==2.0.1
itsdangerous==2.0.1
Werkzeug==2.0.3
psycopg2-binary==2.9.3
opensearch-py==2.0.0
setuptools<58
alembic==1.8.0
img2pdf==0.4.4
appmap>=1.1.0.dev0
boto3==1.34.75
redis==4.6.0
